1001 project1 part3 modifying

+++ src/threads/thread.h
@@ -105,8 +105,8 @@ struct thread
     int priority_old; // right before getting donation
     struct list donation_stack;
     struct thread* child;
-    int nice;
-    int recent_cpu;
+    int64_t nice;
+    int64_t recent_cpu;
 
     /* Owned by thread.c. */
     unsigned magic;                     /* Detects stack overflow. */
